feat: puntos de fidelización dejan de ser monetarios, se canjean por recompensas configurables
Rediseño completo pedido por el usuario: 1 punto ya no equivale a 1 peso.
Se agrega un catálogo de recompensas (LoyaltyReward: nombre, puntos
necesarios, descuento en pesos que otorga) configurable desde el CRM
(Sistema → Administración → Recompensas de Puntos), y tanto el POS como el
checkout del e-commerce se adaptan a elegir una recompensa del catálogo en
vez de escribir un monto libre.

- Puntos pasan a ser enteros (floor() al acreditar, no round() — sin
  fracciones, redondeo hacia abajo). No hizo falta cambiar el tipo de
  columna en la base, alcanza con castear a integer en los modelos.
- LoyaltyService::redeem()/accrue() reescritos: redeem ahora recibe un
  LoyaltyReward (no un monto), debita puntos y devuelve el descuento en
  pesos. accrue() ya no suma directo el amount de un canje anterior (ahora
  son puntos, no pesos) — resuelve los pesos reales vía el
  discount_amount de la recompensa vinculada.
- Nuevo método público validateRedemption() centraliza la validación de
  saldo + tope (max_redemption_percentage) para que POS y checkout no
  dupliquen la misma lógica.
- POS (SaleController::normalizeSalePayments): el split "Puntos" ya no
  acepta un monto — exige loyalty_reward_id, valida y sobrescribe el
  amount contra la recompensa real, antes del chequeo de suma-total.
- Checkout (CatalogController): mismo criterio con loyalty_reward_id en
  vez de loyalty_redeem_amount.
- Saldos y balances existentes NO se migran — el usuario confirmó
  explícitamente que la función todavía no está anunciada a clientes, así
  que se resetean a 0 en el deploy en vez de convertirse 1:1.

+++ artdent-crm/app/Models/LoyaltyAccount.php
@@ -17,7 +17,7 @@ class LoyaltyAccount extends Model
     protected $casts = [
         'customer_id' => 'int',
         'company_id' => 'int',
-        'balance' => 'float',
+        'balance' => 'integer',
     ];
 
     public function customer(): BelongsTo
